+++ packages/lexical-beautiful-mentions/src/BeautifulMentionsPlugin.tsx
@@ -28,10 +28,10 @@ import {
   $isBeautifulMentionNode,
   BeautifulMentionNode,
 } from "./MentionNode";
+import { MenuTextMatch } from "./Menu";
 import TriggerMenuPlugin from "./TriggerMenuPlugin";
 import { TypeaheadMenuPlugin } from "./TypeaheadMenuPlugin";
 import { CAN_USE_DOM } from "./environment";
-import { handleKeydown } from "./handle-keydown";
 import {
   INSERT_MENTION_COMMAND,
   OPEN_MENTIONS_MENU_COMMAND,
@@ -42,11 +42,14 @@ import {
   renameMention,
 } from "./mention-commands";
 import {
-  PUNCTUATION,
-  checkForMentions,
+  DEFAULT_PUNCTUATION,
+  LENGTH_LIMIT,
+  TRIGGERS,
+  VALID_CHARS,
   getCreatableProp,
   getMenuItemLimitProp,
   getSelectionInfo,
+  isWordChar,
 } from "./mention-utils";
 import { useDebounce } from "./useDebounce";
 import { useIsFocused } from "./useIsFocused";
@@ -60,6 +63,113 @@ class MenuOption extends _MenuOption {
     this.value = value;
     this.label = label ?? value;
   }
+}
+
+export function handleKeydown(
+  event: KeyboardEvent,
+  triggers: string[],
+  punctuation: string,
+) {
+  const { key, metaKey, ctrlKey } = event;
+  const simpleKey = key.length === 1;
+  const isTrigger = triggers.some((trigger) => key === trigger);
+  const wordChar = isWordChar(key, triggers, punctuation);
+  const selectionInfo = getSelectionInfo(triggers, punctuation);
+  if (
+    !simpleKey ||
+    (!wordChar && !isTrigger) ||
+    !selectionInfo ||
+    metaKey ||
+    ctrlKey
+  ) {
+    return false;
+  }
+  const {
+    node,
+    offset,
+    isTextNode,
+    textContent,
+    prevNode,
+    nextNode,
+    wordCharAfterCursor,
+    cursorAtStartOfNode,
+    cursorAtEndOfNode,
+  } = selectionInfo;
+  if (isTextNode && cursorAtStartOfNode && $isBeautifulMentionNode(prevNode)) {
+    node.insertBefore($createTextNode(" "));
+    return true;
+  }
+  if (isTextNode && cursorAtEndOfNode && $isBeautifulMentionNode(nextNode)) {
+    node.insertAfter($createTextNode(" "));
+    return true;
+  }
+  if (isTextNode && isTrigger && wordCharAfterCursor) {
+    const content =
+      textContent.substring(0, offset) + " " + textContent.substring(offset);
+    node.setTextContent(content);
+    return true;
+  }
+  if ($isBeautifulMentionNode(node) && nextNode === null) {
+    node.insertAfter($createTextNode(" "));
+    return true;
+  }
+  return false;
+}
+
+// Non-standard series of chars. Each series must be preceded and followed by
+// a valid char.
+const VALID_JOINS = (punctuation: string) =>
+  "(?:" +
+  "\\.[ |$]|" + // E.g. "r. " in "Mr. Smith"
+  "\\s|" + // E.g. " " in "Josh Duck"
+  "[" +
+  punctuation +
+  "]|" + // E.g. "-' in "Salier-Hellendag"
+  ")";
+
+// Regex used to trigger the mention menu.
+function createMentionsRegex(
+  triggers: string[],
+  punctuation: string,
+  allowSpaces: boolean,
+) {
+  return new RegExp(
+    "(^|\\s|\\()(" +
+      TRIGGERS(triggers) +
+      "((?:" +
+      VALID_CHARS(triggers, punctuation) +
+      (allowSpaces ? VALID_JOINS(punctuation) : "") +
+      "){0," +
+      LENGTH_LIMIT +
+      "})" +
+      ")$",
+  );
+}
+
+export function checkForMentions(
+  text: string,
+  triggers: string[],
+  punctuation: string,
+  allowSpaces: boolean,
+): MenuTextMatch | null {
+  const match = createMentionsRegex(triggers, punctuation, allowSpaces).exec(
+    text,
+  );
+  if (match !== null) {
+    // The strategy ignores leading whitespace but we need to know it's
+    // length to add it to the leadOffset
+    const maybeLeadingWhitespace = match[1];
+    const matchingStringWithTrigger = match[2];
+    const matchingString = match[3];
+    if (matchingStringWithTrigger.length >= 1) {
+      return {
+        leadOffset: match.index + maybeLeadingWhitespace.length,
+        matchingString: matchingString,
+        replaceableString: matchingStringWithTrigger,
+      };
+    }
+  }
+  return null;
 }
 
 /**
@@ -78,7 +188,7 @@ export function BeautifulMentionsPlugin(props: BeautifulMentionsPluginProps) {
     showTriggers,
     showMentionsOnDelete,
     mentionEnclosure,
-    punctuation = PUNCTUATION,
+    punctuation = DEFAULT_PUNCTUATION,
   } = props;
   const isEditorFocused = useIsFocused();
   const triggers = useMemo(
