+++ packages/lexical-beautiful-mentions/src/mention-utils.spec.ts
@@ -1,12 +1,16 @@
 import { describe, expect, it } from "vitest";
 import { convertToMentionEntries } from "./mention-converter";
-import { PUNCTUATION } from "./mention-utils";
+import { DEFAULT_PUNCTUATION } from "./mention-utils";
 
 describe("mention-utils", () => {
   it("should get text and mention entries", () => {
     const triggers = ["@", "due:", "#"];
     const text = "Hey @john, the task is #urgent and due:tomorrow";
-    const entries = convertToMentionEntries(text, triggers, PUNCTUATION);
+    const entries = convertToMentionEntries(
+      text,
+      triggers,
+      DEFAULT_PUNCTUATION,
+    );
 
     expect(entries.length).toBe(6);
 
@@ -41,7 +45,11 @@ describe("mention-utils", () => {
   it("should get text and mention entries with multiple mentions", () => {
     const triggers = ["@", "due:", "#"];
     const text = "Hey @john and @jane.";
-    const entries = convertToMentionEntries(text, triggers, PUNCTUATION);
+    const entries = convertToMentionEntries(
+      text,
+      triggers,
+      DEFAULT_PUNCTUATION,
+    );
 
     expect(entries.length).toBe(5);
 
@@ -67,7 +75,11 @@ describe("mention-utils", () => {
   it("should ignore triggers without a value", () => {
     const triggers = ["@", "#"];
     const text = "Hey @ john and # jane.";
-    const entries = convertToMentionEntries(text, triggers, PUNCTUATION);
+    const entries = convertToMentionEntries(
+      text,
+      triggers,
+      DEFAULT_PUNCTUATION,
+    );
     expect(entries.length).toBe(1);
     expect(entries[0].type).toBe("text");
   });
